+++ src/main/kotlin/com/yapp/yapp/running/api/response/RunningStartResponse.kt
@@ -1,4 +1,4 @@
-package com.yapp.yapp.running.response
+package com.yapp.yapp.running.api.response
 
 data class RunningStartResponse(
     val recordId: Long,
